[feat] 음식 상세 조회 구현

+++ backend/befresh/src/main/java/com/a307/befresh/module/domain/container/repository/ContainerRepository.java
@@ -3,6 +3,6 @@ package com.a307.befresh.module.domain.container.repository;
 import com.a307.befresh.module.domain.container.Container;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContainerRepository extends JpaRepository<Container, Integer> {
+public interface ContainerRepository extends JpaRepository<Container, Long> {
     Container findByQrId(long qrId);
 }
